+++ asdf/treeutil.py
@@ -3,12 +3,9 @@ Utility functions for managing tree-like data structures.
 """
 
 import types
-import warnings
 from contextlib import contextmanager
 
 from . import lazy_nodes, tagged
-from .exceptions import AsdfDeprecationWarning, AsdfWarning
-from .util import NotSet
 
 __all__ = ["walk", "iter_tree", "walk_and_modify", "get_children", "is_container", "PendingValue", "RemoveNode"]
 
@@ -221,7 +218,7 @@ class _RemoveNode:
 RemoveNode = _RemoveNode()
 
 
-def walk_and_modify(top, callback, ignore_implicit_conversion=NotSet, postorder=True, _context=None):
+def walk_and_modify(top, callback, postorder=True, _context=None):
     """Modify a tree by walking it with a callback function.  It also has
     the effect of doing a deep copy.
 
@@ -253,24 +250,12 @@ def walk_and_modify(top, callback, ignore_implicit_conversion=NotSet, postorder=
         before their parents.  If `False`, the callable is invoked on the
         parents first.  Defaults to `True`.
 
-    ignore_implicit_conversion : bool
-        DEPRECATED
-        Controls whether warnings should be issued when implicitly converting a
-        given type instance in the tree into a serializable object. The primary
-        case for this is currently ``namedtuple``.
-
-        Defaults to `False`.
-
     Returns
     -------
     tree : object
         The modified tree.
 
     """
-    if ignore_implicit_conversion is NotSet:
-        ignore_implicit_conversion = False
-    else:
-        warnings.warn("ignore_implicit_conversion is deprecated", AsdfDeprecationWarning)
     callback_arity = callback.__code__.co_argcount
     if callback_arity < 1 or callback_arity > 2:
         msg = "Expected callback to accept one or two arguments"
@@ -360,25 +345,18 @@ def walk_and_modify(top, callback, ignore_implicit_conversion=NotSet, postorder=
         # to yield here.
         contents = [_recurse(value, json_id) for value in node]
 
-        try:
-            result = node.__class__(contents)
-            if isinstance(node, tagged.Tagged):
-                result._tag = node._tag
-        except TypeError:
-            # The derived class signature is different, so simply store the
-            # list representing the contents. Currently this is primarily
-            # intended to handle namedtuple and NamedTuple instances.
-            if not ignore_implicit_conversion:
-                warnings.warn(f"Failed to serialize instance of {type(node)}, converting to list instead", AsdfWarning)
-            result = contents
-            warnings.warn("implicit conversion is deprecated. Please instead use a Converter.", AsdfDeprecationWarning)
+        result = node.__class__(contents)
+        if isinstance(node, tagged.Tagged):
+            result._tag = node._tag
 
         return result
 
     def _handle_children(node, json_id):
         if isinstance(node, (dict, lazy_nodes.AsdfDictNode)):
             result = _handle_mapping(node, json_id)
-        elif isinstance(node, tuple):
+        # don't treat namedtuple instances as tuples
+        # see: https://github.com/python/cpython/issues/52044
+        elif isinstance(node, tuple) and not hasattr(node, "_fields"):
             result = _handle_immutable_sequence(node, json_id)
         elif isinstance(node, (list, lazy_nodes.AsdfListNode)):
             result = _handle_mutable_sequence(node, json_id)